+++ src/authres.rs
@@ -13,8 +13,6 @@ use once_cell::sync::Lazy;
 use crate::config::Config;
 use crate::context::Context;
 use crate::headerdef::HeaderDef;
-use crate::mimeparser;
-use crate::mimeparser::ParserErrorExt;
 use crate::tools::time;
 use crate::tools::EmailAddress;
 
@@ -32,23 +30,19 @@ pub(crate) async fn handle_authres(
     mail: &ParsedMail<'_>,
     from: &str,
     message_time: i64,
-) -> mimeparser::ParserResult<DkimResults> {
+) -> Result<DkimResults> {
     let from_domain = match EmailAddress::new(from) {
         Ok(email) => email.domain,
         Err(e) => {
             // This email is invalid, but don't return an error, we still want to
             // add a stub to the database so that it's not downloaded again
-            return Err(anyhow::format_err!("invalid email {}: {:#}", from, e)).map_err_malformed();
+            return Err(anyhow::format_err!("invalid email {}: {:#}", from, e));
         }
     };
 
     let authres = parse_authres_headers(&mail.get_headers(), &from_domain);
-    update_authservid_candidates(context, &authres)
-        .await
-        .map_err_sql()?;
-    compute_dkim_results(context, authres, &from_domain, message_time)
-        .await
-        .map_err_sql()
+    update_authservid_candidates(context, &authres).await?;
+    compute_dkim_results(context, authres, &from_domain, message_time).await
 }
 
 #[derive(Default, Debug)]