+++ src/tests/aeap.rs
@@ -341,9 +341,8 @@ async fn mark_as_verified(this: &TestContext, other: &TestContext) {
 
     peerstate.verified_key = peerstate.public_key.clone();
     peerstate.verified_key_fingerprint = peerstate.public_key_fingerprint.clone();
-    peerstate.to_save = Some(peerstate::ToSave::All);
 
-    peerstate.save_to_db(&this.sql, false).await.unwrap();
+    peerstate.save_to_db(&this.sql).await.unwrap();
 }
 
 async fn get_last_info_msg(t: &TestContext, chat_id: ChatId) -> Option<Message> {