+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 ## Unreleased
 
 ### Changes
+- Refactor: Remove the remaining AsRef<str> #3669
+- Small speedup #3780
+
+### API-Changes
+- Add Python API to send reactions #3762
+- jsonrpc: add message errors to MessageObject #3788
+
+### Fixes
+- Make sure malformed messsages will never block receiving further messages anymore #3771
+- strip leading/trailing whitespace from "Chat-Group-Name{,-Changed}:" headers content #3650
+- Assume all Thunderbird users prefer encryption #3774
+- refactor peerstate handling to ensure no duplicate peerstates #3776
+
+
+## 1.102.0
+
+### Changes
 
 - If an email has multiple From addresses, handle this as if there was
   no From address, to prevent from forgery attacks. Also, improve